% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdqa_assert.R
\name{sdqa_assert_geom_simple}
\alias{sdqa_assert_geom_simple}
\title{Assert that all geometries in a spatial object are simple}
\usage{
sdqa_assert_geom_simple(x)
}
\arguments{
\item{x}{A spatial object. Supported classes: [`sf`][sf::sf],
[`SpatVector`][terra::SpatVector], or `duckspatial_df`.
[`SpatRaster`][terra::SpatRaster] is not supported.}
}
\value{
Invisibly returns `TRUE`. Throws an error if any geometry is
  non-simple.
}
\description{
Checks whether every geometry in `x` is simple (i.e. does not
self-intersect). Throws an error listing the row indices of any non-simple
geometries.
}
\examples{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)

sdqa_assert_geom_simple(nc)

\dontrun{
non_simple <- st_sf(
  geometry = st_sfc(
    st_linestring(matrix(
      c(0, 0, 1, 1, 0, 1, 1, 0), ncol = 2, byrow = TRUE
    )),
    crs = 4326
  )
)
sdqa_assert_geom_simple(non_simple)
}
}
